make printer and input singleton, ported some ncurse.h fn or marco to ncurses,cpp

+++ makefile
@@ -3,16 +3,16 @@ CXXFLAGS	= -O3
 LDFLAGS		= -lncurses
 
 test: CXXFLAGS = -g
-test: test.o input.o printer.o ncurses.o
+test: test.o input.o printer.o ncurses.o game.o board.o
 	$(CXX) $(CXXFLAGS) $^ $(LDFLAGS) -o $@
 
-(*.o) : $(*.cpp)
+(*.o) : $(*.cpp) $(*.hpp) util.h
 	$(CXX) $(CXXFLAGS) -c $<
 
 clean_exec : 
-	rm test
+	if (ls ./test); then rm test; fi
 
 clean_obj : 
-	rm *.o
+	if (ls ./*.o); then rm *.o; fi
 
 clean : clean_exec clean_obj
